pyproject.toml
[tool.poetry]
name = "htmx_demo"
version = "0.1.0"
description = "Django DEMO project for htmx live user feedback"
authors = ["JensDiemer <git@jensdiemer.de>"]
homepage = "https://github.com/jedie/htmx_demo"
license = "GPL v3 or above"

[tool.poetry.dependencies]
python = ">=3.9,<4.0.0"
Django = "*"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
